+++ SqExpress/SqlExport/Internal/TSqlBuilder.cs
@@ -274,6 +274,15 @@ namespace SqExpress.SqlExport.Internal
             return true;
         }
 
+        public override bool VisitExprTypeFixSizeByteArray(ExprTypeFixSizeByteArray exprTypeFixSizeByteArray, IExpr? arg)
+        {
+            this.Builder.Append("binary(");
+            this.Builder.Append(exprTypeFixSizeByteArray.Size.ToString());
+            this.Builder.Append(')');
+
+            return true;
+        }
+
         public override bool VisitExprTypeInt16(ExprTypeInt16 exprTypeInt16, IExpr? parent)
         {
             this.Builder.Append("smallint");
@@ -380,6 +389,24 @@ namespace SqExpress.SqlExport.Internal
                     this.Builder.Append("(MAX)");
                 }
             }
+
+            return true;
+        }
+
+        public override bool VisitExprTypeFixSizeString(ExprTypeFixSizeString exprTypeFixSizeString, IExpr? arg)
+        {
+            if (exprTypeFixSizeString.IsUnicode)
+            {
+                this.AppendName("nchar");
+            }
+            else
+            {
+                this.AppendName("char");
+            }
+
+            this.Builder.Append('(');
+            this.Builder.Append(exprTypeFixSizeString.Size.ToString());
+            this.Builder.Append(')');
 
             return true;
         }