+++ SqExpress/SyntaxTreeOperations/ExprDeserializer.cs
@@ -143,6 +143,8 @@ namespace SqExpress.SyntaxTreeOperations
                 case "TypeDateTime": return new ExprTypeDateTime(isDate: ReadBoolean(rootElement, reader, "IsDate"));
                 case "TypeDecimal": return new ExprTypeDecimal(precisionScale: ReadNullableDecimalPrecisionScale(rootElement, reader, "PrecisionScale"));
                 case "TypeDouble": return ExprTypeDouble.Instance;
+                case "TypeFixSizeByteArray": return new ExprTypeFixSizeByteArray(size: ReadInt32(rootElement, reader, "Size"));
+                case "TypeFixSizeString": return new ExprTypeFixSizeString(size: ReadInt32(rootElement, reader, "Size"), isUnicode: ReadBoolean(rootElement, reader, "IsUnicode"));
                 case "TypeGuid": return ExprTypeGuid.Instance;
                 case "TypeInt16": return ExprTypeInt16.Instance;
                 case "TypeInt32": return ExprTypeInt32.Instance;